requirements.txt
# Trading Platform Requirements
# ================================

# Core dependencies (required)
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0.0
joblib>=1.3.0

# Machine Learning (required for ML features)
scikit-learn>=1.3.0

# Machine Learning (optional - enhanced performance)
# xgboost>=2.0.0
# lightgbm>=4.0.0

# Deep Learning (optional - for LSTM strategies)
# tensorflow>=2.15.0

# Technical Analysis (recommended)
ta>=0.10.0

# GUI (required for graphical interface)
PyQt6>=6.6.0

# Visualization (required for charts)
matplotlib>=3.8.0

# Trading (required - Windows only)
# Install separately: pip install MetaTrader5
# MetaTrader5>=5.0.0

# Database (optional - for production)
# sqlalchemy>=2.0.0
# psycopg2-binary>=2.9.0
# redis>=5.0.0
# influxdb-client>=1.39.0

# Data Sources (optional)
# yfinance>=0.2.0
# ccxt>=4.0.0

# Optimization (optional)
# deap>=1.4.0
# bayesian-optimization>=1.4.0

# Additional Visualization (optional)
# plotly>=5.17.0
# seaborn>=0.13.0

# Performance (optional)
# numba>=0.58.0
# psutil>=5.9.0

# Utilities (optional)
# tqdm>=4.66.0

# Development (optional)
# pytest>=7.4.0
# black>=23.9.0
# flake8>=6.1.0


# =================================
# INSTALLATION GUIDE
# =================================
# 
# Basic installation (GUI + Core features):
#   pip install pandas numpy pyyaml joblib scikit-learn ta PyQt6 matplotlib
#
# Full installation with ML:
#   pip install pandas numpy pyyaml joblib scikit-learn ta PyQt6 matplotlib xgboost lightgbm
#
# MetaTrader 5 (Windows only):
#   pip install MetaTrader5
#
# Deep Learning (optional):
#   pip install tensorflow
